fix(#4900): keep routine loader below giant threshold
Extract discovery helpers so the loader remains below the production-line registry limit and Script checks can regenerate inventory cleanly.
